Create College data base managementsystem using python
this project is based on DBMS . We used pysql and some small python functions to build this code. everything went well till we found that the data base is not connected to the code we have written and our code is not working that well. we will still try to improve this project. so this project is still in progress for us.